Apply self-suggestions
